# Optimize Ruby code