Integer#times is faster than using Ranges, see pull request #23 for details.