@PEZ 2024-12-21
Thanks! 🙏  

It would be interesting to know how much faster. 😄 Can you post benchmark runs from before and after?