@gettalong 2024-12-22
Using Ruby 3.4 (`ruby 3.4.0dev (2024-11-18T15:04:50Z master 4a7ac694e5) +PRISM [x86_64-linux]`) the non-YJIT run gains a bit more than 5%. The YJIT run, however, gains around 32%!

Before:

~~~
Checking Ruby YJIT
Check passed
Benchmarking Ruby YJIT
Benchmark 1: ruby --yjit ./ruby/code.rb aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):     500.8 ms ±   2.1 ms    [User: 462.0 ms, System: 38.9 ms]
  Range (min … max):   498.4 ms … 502.4 ms    3 runs


Checking Ruby
Check passed
Benchmarking Ruby
Benchmark 1: ruby ./ruby/code.rb aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):      1.016 s ±  0.012 s    [User: 0.976 s, System: 0.040 s]
  Range (min … max):    1.008 s …  1.029 s    3 runs
~~~

After:

~~~
Checking Ruby YJIT
Check passed
Benchmarking Ruby YJIT
Benchmark 1: ruby --yjit ./ruby/code.rb aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):     309.2 ms ±   2.0 ms    [User: 270.3 ms, System: 39.2 ms]
  Range (min … max):   307.8 ms … 311.5 ms    3 runs


Checking Ruby
Check passed
Benchmarking Ruby
Benchmark 1: ruby ./ruby/code.rb aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa ...
  Time (mean ± σ):     959.6 ms ±   9.0 ms    [User: 921.1 ms, System: 38.9 ms]
  Range (min … max):   952.7 ms … 969.8 ms    3 runs
~~~